Parse test cases from json file
